Refactor base::Timer to not call TimeTicks::Now for task with 0 delay.
This CL remove the calls to TimeTicks::Now when a timer is called with a
delay null or negative. The reason is that TimeTicks::Now is costly on
Android and is not useful in that use case.

R=mark@chromium.org
BUG=315070

Review URL: https://codereview.chromium.org/105003002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d04519357452f010ebc4c8e237333045058da21e