message-pump: Change how MessagePumpDispatchers process events.
This patch allows a MessagePumpDispatcher to invoke the default handling of the message
in the message-pump by setting the DISPATCH_DEFAULT flag on the return value from the
::Dispatch() callback. As before, the dispatcher can still request a termination of the
current message-loop, but setting the DISPATCH_QUIT flag on the return value (instead of
returning false). The callback can set both flags, in which case the message-loop will
perform the default action, and also terminate the current loop.

As a result of this, we can remove duplicate code, namely aura::DispatcherWin, and
aura::Env::GetDispatcher(), that performs the same action as the default message pump.

R=darin@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/152083004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 93805f36f94bd5f824cff64e2f61e24a2df16d6d