Shift by the right number of bytes when grabbing the high 32 bits.
When size_t is not 64 bits (8 bytes) but instead N < 8 bytes, we
want to grab the top N bytes from the 64-bit result.

This code was meant to do this:
hash64 >> (sizeof(uint64) - sizeof(std::size_t))

But this translates to:
hash64 >> (8 - N)

And 8-N is the number of bytes we want to shift, not the number of
bits, where as operator>> excepts a number of bits.

Fix this to correct the algorithm for platforms with small size_t.

R=ajwong
BUG=

Review URL: https://codereview.chromium.org/156103003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 51421855d70a15031f7046efd771db6d6f453415