POSIX: CHECK() that file_util::ScopedFD fulfills promise.
CHECK() that file_util::ScopedFD will actually close the file descriptors.
There are security implications to not doing so, and logging an error is not
enough: file descriptor are security capabilities. Failing to close them is a
failure to revoke access to certain resources, which is heavily relied on in the
code base.

This CL also adds unit tests to file_util::ScopedFD.

Review URL: https://codereview.chromium.org/183953004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fb73aed94a95f568a846644cd60021b2c5569274