Handle errors properly in FileUtilProxy::CreateTemporary.
If creating the temporary file failed, don't try to open it. If we can't open
it for some reason, delete the temporary so the caller can assume strong
exception safety and not have to delete the file on error.

BUG=none

Review URL: https://codereview.chromium.org/183003009

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b4d3ed61a60f824caf19a92a00bd7f10ea53052c