Add a wrapper script which modifies the .isolate on the fly.
This tool will ultimately process the .ninja files, in the meantime it archives
all the DLLs. It's cheezy but this will unblock the component build short term.
This is very crude but this will get folks going. This means once it's done, we
can remove all the .isolate representing the dynamic libraries.

Remove swarming_client/ as a dependency in the .isolate files, it's not needed
anymore.

R=vadimsh@chromium.org
BUG=333473

Review URL: https://codereview.chromium.org/196283016

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc64721761637761afcbc166ea175ec338c463f8