base: Do not allow MessagePumpObservers to consume events.
There is currently a single message-pump observer that can consume an event,
and it does so only when the num-lock key is pressed on Chrome OS. But it
doesn't need to consume the event, because no other dispatcher/observer cares
about this event. So remove a MessagePumpObserver's ability to consume an
event. This makes the code cleaner, and easier to understand.

BUG=354062
R=darin@chromium.org, derat@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/223483002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 030a0491b1d2e0bda3933e0d1f4a6a90c0b06853