Remove CreatePlatformFile from file_util_proxy
Most of file_util_proxy is going to be deleted soon (users have to migrate to
FileProxy) so this CL is quite superficial. It is intended to remove the last
uses of CreatePlatformFile so that the method can be removed and no new code
can ignore base::File.

BUG=322664

Review URL: https://codereview.chromium.org/227433006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4735522f8b30a06dce580b770f79020adc730352