ozone: Remove the libevent message-pump.
The ozone message-pump is no longer needed. So remove that, and use the default
message-pump based on libevent instead. Also, add a convenience method for
adding a watcher on a file-descriptor to the message-pump from the message-loop.

BUG=354062
R=rjkroege@chromium.org, sky@chromium.org
TBR=darin@chromium.org

Review URL: https://codereview.chromium.org/231643005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ef1a61b3ca42b158068c27e722630d881eb5fa1e