Allow creating SharedMemory that can't be shared read-only.
Sharing a SharedMemory read-only requires that an extra fd be opened when it was created, which can cause problems on systems where fds are in short supply. To avoid that, only create that fd when it's necessary.

BUG=362603

Review URL: https://codereview.chromium.org/241443002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: add1a2674280222b74f3f31b58da43001110a1c0