Fix mismatched bookmark favicons on Android 4.2.1.
Due to a platform bug in Android 4.2.1 (JB MR1),
TextView.setCompoundDrawablesRelative() is a no-op if the view has ever
been measured. As a workaround, we use setCompoundDrawables() directly.

This bug caused the favicons on the bookmarks to appear next to the
wrong bookmark (e.g. the Gmail favicon might appear next to the Twitter
bookmark). This is now fixed.

BUG=368196

Review URL: https://codereview.chromium.org/263473005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d2b739de6e881f7d2091fe20784206f2e60e663b