Create copy of context record before passing it to StackWalk64
StackWalk64 may modify EXCEPTION_POINTERS struct that is passed
to it from constructor of StackTrace.
In the case of StackDumpExceptionFilter the struct that is passed
to constructor of StackTrace may be later passed to other (chained)
exception filter(s), so it shouldn't be modified.

BUG=

Review URL: https://codereview.chromium.org/267463003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe7a12a3fe6a32fafe8aa01305fd965262d69755