Store a stacktrace of PrefService destruction in PrefChangeRegistrar
Store a stacktrace of PrefService destruction in PrefChangeRegistrar to debug a
race condition: ~PrefChangeRegistrar causes a crash. The current hypothesis is
that this is caused because the PrefChangeRegistrar is registered to a Profile
that is already destroyed. This code stores stacktrace when the Profile's
PrefService is destroyed to see whether this is correct.

BUG=373435

Review URL: https://codereview.chromium.org/290083006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 02a7dfa41a288ecac3f14324b5009d19fb208430