Makes MessagePumpGlib thread DCHECK work when pump is destroyed
For tests I'm creating a MessagePumpGlib on a thread, destroying it,
then creating another MessagePumpGlib on another thread. The DCHECK
fires in this case because it only remembers the first thread. I made
it so the thread id is cleared when the MessagePumpGlib is destroyed.

BUG=none
TEST=none
R=sadrul@chromium.org
TBR=darin@chromium.org

Review URL: https://codereview.chromium.org/289043010

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3154c3cb41b75321bc746eda5ccdfbd54180affc