Add java_cpp_template template
This is the GN version of build/java_cpp_template.gypi.

It defines a template that wraps an action_foreach that generates .java
files using the host C preprocessor.

The major difference in the GN version is that it takes all the
generated java files and zips them together in a single .srcjar. When
such a target is included in the srcjar_deps of a java library, the
.java files in the .srcjar will be treated much like files listed in
that libraries java_sources (particularly they will be compiled and
included in the .jar/.dex).

Depends on https://crrev.com/264773014/

BUG=359249

Review URL: https://codereview.chromium.org/264923007

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6fad603bfebe6314ecaa4d74a3ffdca430eeb8a6