Another try at fixing --enable-logging on Windows
Per details logged on connect bug and crbug, when linked as a Windows
subsystem binary, the _fileno() for stdin/stdout/stderr are incorrectly set
to 0,1,2 rather than all -2 (because there's no console attached). This 
was broken between VS2010->VS2012 and is still broken in the latest VS2013
update.

So, instead of only checking _fileno, we also check the underlying HANDLE
to confirm that it's valid for that stream before aborting.

BUG=358267
BUG=https://connect.microsoft.com/VisualStudio/feedback/details/785119/#tabs

Review URL: https://codereview.chromium.org/324493003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 36fc75d4515eac399277c308b06ea90869af1f3d