Password bubble: Ensure that the bubble stays open for at least 1 second.
Upon navigation, we throw away everything in the password bubble. This
is intentional, but a poor experience on sites that do strange things on
login. Like `accounts.google.com` for example.

This patch ensures that we don't clear the bubble state if a navigation
occurs within a second. A future patch could extend this to not clearing
the bubble state if navigations occur within the same eTLD+1.

BUG=383296
TBR=awong@chromium.org

Review URL: https://codereview.chromium.org/338103003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aaf6cce59e126b77ad17f2e8452c83ae56e0111f